refactor(settings): extract password-rule memos into usePasswordRules
Move newPasswordRules/confirmPasswordRules verbatim into
_hooks/use-password-rules.ts; hook owns its own Auth.passwordRules +
Settings translations and takes the two watched field values. Drops the
now-unused PASSWORD_RULES and useMemo imports. Behavior-identical;
page.tsx 1591->1572 LOC.
